# Optimize distributed model sync